Use kebab case for file names
