+++ src/async-check.ts
