+++ src/index.ts
@@ -2,20 +2,20 @@ export { default as Result } from "./result";
 export { default as ok, OkResult } from "./ok";
 export { default as error, ErrorResult } from "./error";
 export { default as Check } from "./check";
-export { default as AsyncCheck } from "./asyncCheck";
-export { default as MaybeAsyncCheck } from "./maybeAsyncCheck";
+export { default as AsyncCheck } from "./async-check";
+export { default as MaybeAsyncCheck } from "./maybe-async-check";
 
 export { default as pass } from "./pass";
 export { default as fail } from "./fail";
 export { default as test } from "./test";
-export { default as testAsync } from "./testAsync";
+export { default as testAsync } from "./test-async";
 export { default as transform } from "./transform";
-export { default as transformAsync } from "./transformAsync";
+export { default as transformAsync } from "./transform-async";
 
 export { default as is } from "./is";
 export { default as null } from "./null";
 export { default as undefined } from "./undefined";
-export { default as NaN } from "./NaN";
+export { default as NaN } from "./not-a-number";
 export { default as boolean } from "./boolean";
 export { default as number } from "./number";
 export { default as integer } from "./integer";
@@ -25,10 +25,10 @@ export { default as date } from "./date";
 export { default as object } from "./object";
 export { default as array } from "./array";
 
-export { default as toBoolean } from "./toBoolean";
-export { default as toNumber } from "./toNumber";
-export { default as toString } from "./toString";
-export { default as toDate } from "./toDate";
+export { default as toBoolean } from "./to-boolean";
+export { default as toNumber } from "./to-number";
+export { default as toString } from "./to-string";
+export { default as toDate } from "./to-date";
 
 export { default as floor } from "./floor";
 export { default as ceil } from "./ceil";
@@ -36,34 +36,34 @@ export { default as round } from "./round";
 export { default as truncate } from "./truncate";
 export { default as min } from "./min";
 export { default as max } from "./max";
-export { default as lessThan } from "./lessThan";
-export { default as moreThan } from "./moreThan";
+export { default as lessThan } from "./less-than";
+export { default as moreThan } from "./more-than";
 
 export { default as trim } from "./trim";
-export { default as toLower } from "./toLower";
-export { default as toUpper } from "./toUpper";
+export { default as toLower } from "./to-lower";
+export { default as toUpper } from "./to-upper";
 export { default as pattern } from "./pattern";
 
-export { default as minLength } from "./minLength";
-export { default as maxLength } from "./maxLength";
+export { default as minLength } from "./min-length";
+export { default as maxLength } from "./max-length";
 
 export { default as not } from "./not";
-export { default as notAsync } from "./notAsync";
+export { default as notAsync } from "./not-async";
 export { default as chain } from "./chain";
-export { default as chainAsync } from "./chainAsync";
-export { default as oneOf } from "./oneOf";
-export { default as oneOfAsync } from "./oneOfAsync";
+export { default as chainAsync } from "./chain-async";
+export { default as oneOf } from "./one-of";
+export { default as oneOfAsync } from "./one-of-async";
 
 export { default as optional } from "./optional";
-export { default as optionalAsync } from "./optionalAsync";
+export { default as optionalAsync } from "./optional-async";
 export { default as nullable } from "./nullable";
-export { default as nullableAsync } from "./nullableAsync";
+export { default as nullableAsync } from "./nullable-async";
 
 export { default as items } from "./items";
-export { default as itemsAsync } from "./itemsAsync";
+export { default as itemsAsync } from "./items-async";
 export { default as shape } from "./shape";
-export { default as shapeAsync } from "./shapeAsync";
+export { default as shapeAsync } from "./shape-async";
 export { default as tuple } from "./tuple";
-export { default as tupleAsync } from "./tupleAsync";
+export { default as tupleAsync } from "./tuple-async";
 export { default as entries } from "./entries";
-export { default as entriesAsync } from "./entriesAsync";
+export { default as entriesAsync } from "./entries-async";
