+++ src/maybe-async-check.ts
