+++ src/optional-async.ts
@@ -1,14 +1,14 @@
-import AsyncCheck from "./asyncCheck";
-import MaybeAsyncCheck from "./maybeAsyncCheck";
+import AsyncCheck from "./async-check";
+import MaybeAsyncCheck from "./maybe-async-check";
 import ok from "./ok";
 
 /**
- * Asynchronous version of [[`nullable`]].
+ * Asynchronous version of [[`optional`]].
  *
  * ```js
- * const check = nullableAsync(testAsync(async value => value === 42));
+ * const check = optionalAsync(testAsync(async value => value === 42));
  *
- * await check(null);
+ * await check(undefined);
  * // => { isOk: true, ... }
  * await check(42);
  * // => { isOk: true, ... }
@@ -19,12 +19,12 @@ import ok from "./ok";
  * @param check The async check function to run when the value is present.
  * @returns An async check function.
  */
-export default function nullableAsync<I, O, A extends unknown[]>(
+export default function optionalAsync<I, O, A extends unknown[]>(
   check: MaybeAsyncCheck<I, O, A>,
-): AsyncCheck<null | I, null | O, A> {
+): AsyncCheck<undefined | I, undefined | O, A> {
   return async (value, ...args) => {
-    if (value === null) {
-      return ok(null);
+    if (value === undefined) {
+      return ok(undefined);
     } else {
       return await check(value, ...args);
     }
