+++ src/to-boolean.ts
