+++ test/entries-async.test.ts
