+++ test/not-async.test.ts
